// sanity.types.ts
export type Skill = {
  _id: string;
  _type: "skill";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  category?: "frontend" | "backend" | "ai-ml" | "database" | "mobile" | "testing" | "design" | "tools" | "other";
  proficiency?: "beginner" | "intermediate" | "advanced" | "expert";
  percentage?: number;
  yearsOfExperience?: number;
  color?: string;
};